# internal/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LLM Gateway 统一大模型 API",
    "version": "1.0.0",
    "description": "OpenAI 兼容的统一大模型接入接口。业务只需携带 `model` 调用统一 API，网关负责厂商匹配、按 API Key 命中调度策略、跨厂商容灾、Key/模型/部门三级 TPM·QPS 限流与预算管控。\n\n可直接使用 OpenAI 官方 SDK，只需把 `base_url` 指向网关、`api_key` 换成网关分发的 Key。"
  },
  "servers": [{ "url": "/v1" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/models": {
      "get": {
        "tags": ["模型"],
        "operationId": "listModels",
        "summary": "列出可用模型",
        "description": "返回模型市场中已上架的模型。",
        "responses": {
          "200": {
            "description": "成功",
            "content": { "application/json": {
              "schema": { "$ref": "#/components/schemas/ModelList" },
              "example": { "object": "list", "data": [{ "id": "deepseek-r1", "object": "model", "owned_by": "aliyun" }] }
            } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "网关分发的 API Key，形如 `sk-xxxxxxxx`。在控制台「API Key · Key 申请」提交，审批后获得（密钥仅展示一次）。" }
    },
...
    "schemas": {
...
      "ModelList": {
        "type": "object",
        "properties": {
          "object": { "type": "string", "example": "list" },
          "data": { "type": "array", "items": { "type": "object", "properties": {
            "id": { "type": "string" }, "object": { "type": "string" }, "owned_by": { "type": "string", "description": "厂商编码" }
          } } }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "object", "properties": {
            "message": { "type": "string", "description": "可读的错误说明" },
            "type": { "type": "string", "description": "错误大类" },
            "code": { "type": "string", "description": "机器可读的错误码" },
            "param": { "type": "string" }
          } }
        }
      }
    },
    "responses": {
...
      "Unauthorized": { "description": "缺少、无效或未生效的 API Key（`missing_api_key` / `invalid_api_key` / `inactive_api_key`）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": { "message": "Invalid API key", "type": "invalid_request_error", "code": "invalid_api_key" } } } } },
...
      "Forbidden": { "description": "调用方 IP 不在该 Key 的 IP 白名单内（`ip_not_allowed`）、请求的模型不在该 Key 的可用模型中（`model_not_allowed`），或 Key 所属部门已停用（`department_disabled`）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    }
  }
}